+++ tests/bundleScore.ts
@@ -2,7 +2,7 @@
 import fs from "fs";
 
 import "../env.js";
-import {markScore} from "../backend/scoreMaker";
+import {markScore, markScoreParallelly} from "../backend/scoreMaker";
 import loadLilyParser from "../backend/loadLilyParserNode";
 import LogRecorder from "../inc/logRecorder";
 
@@ -14,8 +14,10 @@ const main = async (...args) => {
 	const lilyParser = await loadLilyParser();
 
 	const ly = fs.readFileSync(args[0]).toString();
+	const t0 = Date.now();
 	const result = await markScore(ly, lilyParser, {logger});
-	console.log("result:", result, logger);
+	console.log("cost:", Date.now() - t0);
+	//console.log("result:", result, logger);
 
 	const outputPath = args[1];
 	if (outputPath) {